refactor: 完成属性系统、标签系统大重构
这是一次大规模的代码重构，主要变更包括：
1.  重构属性系统：使用字符串ID替代枚举作为属性标识，统一所有角色属性配置格式
2.  替换旧Buff系统为Modifier系统，移除大量废弃代码和兼容逻辑
3.  重命名大量GameplayTag常量，使用更规范的命名空间格式
4.  调整AI目标选择逻辑，修复排序错误
5.  简化路径查找逻辑，优化地形成本计算标签匹配
6.  统一测试代码中的属性访问方式，移除废弃的AttributeKind枚举使用
7.  修复多个调试查看器的兼容性问题
8.  更新所有角色配置文件，适配新的属性配置格式

+++ src/core/ai/targeting.rs
@@ -9,9 +9,9 @@ pub(crate) struct UnitSnapshot {
     pub entity: Entity,
     pub faction: Faction,
     pub coord: IVec2,
-    pub atk: f32,
-    pub hp: f32,
-    pub max_hp: f32,
+    pub atk: i32,
+    pub hp: i32,
+    pub max_hp: i32,
     pub mov: u32,
     pub attack_range: u32,
     pub acted: bool,
@@ -51,9 +51,9 @@ mod tests {
         entity: Entity,
         faction: Faction,
         coord: IVec2,
-        atk: f32,
-        hp: f32,
-        max_hp: f32,
+        atk: i32,
+        hp: i32,
+        max_hp: i32,
     ) -> UnitSnapshot {
         UnitSnapshot {
             entity,
@@ -78,17 +78,17 @@ mod tests {
             Entity::PLACEHOLDER,
             Faction::Player,
             IVec2::new(3, 0),
-            10.0,
-            20.0,
-            20.0,
+            10,
+            20,
+            20,
         );
         let p2 = make_snapshot(
             Entity::PLACEHOLDER,
             Faction::Player,
             IVec2::new(10, 10),
-            10.0,
-            20.0,
-            20.0,
+            10,
+            20,
+            20,
         );
         let snapshots = vec![p1, p2];
         let selector = NearestTarget;
@@ -102,17 +102,17 @@ mod tests {
             Entity::PLACEHOLDER,
             Faction::Player,
             IVec2::new(3, 0),
-            10.0,
-            5.0,
-            20.0,
+            10,
+            5,
+            20,
         );
         let p2 = make_snapshot(
             Entity::PLACEHOLDER,
             Faction::Player,
             IVec2::new(5, 0),
-            10.0,
-            18.0,
-            20.0,
+            10,
+            18,
+            20,
         );
         let snapshots = vec![p1, p2];
         let selector = WeakestTarget;
@@ -126,17 +126,17 @@ mod tests {
             Entity::PLACEHOLDER,
             Faction::Player,
             IVec2::new(3, 0),
-            5.0,
-            20.0,
-            20.0,
+            5,
+            20,
+            20,
         );
         let p2 = make_snapshot(
             Entity::PLACEHOLDER,
             Faction::Player,
             IVec2::new(5, 0),
-            15.0,
-            20.0,
-            20.0,
+            15,
+            20,
+            20,
         );
         let snapshots = vec![p1, p2];
         let selector = MostDangerousTarget;
@@ -150,17 +150,17 @@ mod tests {
             Entity::PLACEHOLDER,
             Faction::Player,
             IVec2::new(3, 0),
-            10.0,
-            2.0,
-            20.0,
+            10,
+            2,
+            20,
         );
         let p2 = make_snapshot(
             Entity::PLACEHOLDER,
             Faction::Player,
             IVec2::new(5, 0),
-            10.0,
-            15.0,
-            20.0,
+            10,
+            15,
+            20,
         );
         let snapshots = vec![p1, p2];
         let selector = LowestHpPercentTarget;
@@ -183,17 +183,17 @@ mod tests {
             Entity::PLACEHOLDER,
             Faction::Player,
             IVec2::new(3, 0),
-            10.0,
-            5.0,
-            20.0,
+            10,
+            5,
+            20,
         );
         let p2 = make_snapshot(
             Entity::PLACEHOLDER,
             Faction::Player,
             IVec2::new(5, 0),
-            15.0,
-            18.0,
-            20.0,
+            15,
+            18,
+            20,
         );
         let snapshots = vec![p1, p2];
